// apps/legacy-hono-api/src/v1/schema.ts
export const sdkPaywallResponseSchema = z
  .object({
    paywallId: z.string(),
    paywallProducts: z.array(
      z.object({
        paywallProductId: z.string(),
        productId: z.string(),
        displayName: z.string(),
        price: z.number().nullable(),
        nativePurchaseAvailable: z.boolean(),
        webCheckoutAvailable: z.boolean(),
        webCheckoutPaymentProviderConfigurationProductId: z.string().nullable()
      })
    )
  })
  .meta({
    ref: 'SdkPaywall'
  });
// ...
export const sdkGetConfigurationResponseSchema = z.object({
  paywalls: z.array(
    z.object({
      paywallId: z.string(),
      paywallProducts: z.array(
        z.object({
          paywallProductId: z.string(),
          productId: z.string(),
          displayName: z.string(),
          nativePaymentProviderConfigurationProductId: z.string().nullable(),
          defaultWebCheckoutPaymentProviderConfigurationProductId: z
            .string()
            .nullable(),
          paymentProviderConfigurationProducts: z.array(
            z.object({
              paymentProviderConfigurationProductId: z.string(),
              paymentProviderConfigurationId: z.string(),
              configuration: z.record(z.string(), z.any())
            })
          )
        })
      )
    })
  ),
  paywallLocations: z.array(
    z.object({
      paywallLocationId: z.string(),
      slug: z.string()
    })
  ),
  placements: z.array(
    z.object({
      paywallId: z.string(),
      paywallLocationId: z.string()
    })
  ),
  paymentProviderConfigurations: z.array(
    z.object({
      paymentProviderConfigurationId: z.string(),
      providerId: z.string()
    })
  )
});